Updated video
Hi Kevin,

Could you please use the attached text for the next 3,000 Strong email. Please put the video image in the same place of the email as you did last time. Just change the text. Here is the URL for the new video:

https://www.youtube.com/watch?v=XK93kVGLZ6Y&feature=player_detailpage

Please embed this video in the 3,000 Strong site and make it the first to appear. Everything we’ve changed on the site looks good in the proofs that you’ve sent.

Please send me a proof of the email when it is ready.

Many, many thanks

S