Bugfix
One minor detail, on the first mention of $200,000.00 we use the ‘.00’ at the very end. Later, we don’t use it. I corrected this in some other versions but it made it to you—would you mind dropping that .00 off of the dollar amount in the second sentence of the first paragraph?